[module2-task1] rename camelCase filenames to kebab-case
