+++ src/components/film-details-popup.js
