+++ src/components/index.js
@@ -1,6 +1,6 @@
 export * from './statistics';
 export * from './navigation';
-export * from './userLevel';
+export * from './user-level';
 export * from './sort';
 export * from './films';
-export * from './filmDetailsPopup';
+export * from './film-details-popup';
